Move MAX_CONFIG_SIZE to constants #617E
