+++ ui/app/src/components/GlobalConfigManagement/GlobalConfigManagement.tsx
@@ -33,7 +33,8 @@ import { useHistory } from 'react-router';
 import ConfirmDialog from '../ConfirmDialog/ConfirmDialog';
 import { useMount } from '../../hooks/useMount';
 import Paper from '@mui/material/Paper';
-import { getContentSizeInBytes, MAX_CONFIG_SIZE } from '../../utils/content';
+import { getContentSizeInBytes } from '../../utils/content';
+import { MAX_CONFIG_SIZE } from '../../utils/constants';
 
 const translations = defineMessages({
   configSaved: {
